create class ConfigurationUtils rename util package to utils
