+++ src/main/java/config/ConfigurationUtils.java
@@ -1,0 +1,23 @@
+package config;
+
+import java.io.FileInputStream;
+import java.io.FileNotFoundException;
+import java.io.IOException;
+import java.util.Properties;
+
+public class ConfigurationUtils {
+    // last properties files override first
+    public static Properties loadProperties(String... propertiesPaths) {
+        var res = new Properties();
+        try {
+            for (String path : propertiesPaths) {
+                res.load(new FileInputStream(path));
+            }
+        } catch (FileNotFoundException e) {
+            throw new ConfigurationFailException("Configuration files are not found", e);
+        } catch (IOException e) {
+            throw new ConfigurationFailException("Configuration files are not loaded", e);
+        }
+        return res;
+    }
+}
