+++ src/main/java/extractor/DefaultExtractor.java
@@ -1,17 +1,7 @@
 package extractor;
 
-import org.jsoup.Jsoup;
-import org.jsoup.nodes.Document;
-import org.jsoup.nodes.Element;
-import util.HTML;
-import util.Link;
+import utils.HTML;
 
-import java.io.ByteArrayInputStream;
-import java.io.File;
-import java.io.IOException;
-import java.net.URL;
-import java.util.Arrays;
-import java.util.HashSet;
 import java.util.Set;
