+++ src/main/java/utils/CSVParser.java
@@ -1,6 +1,4 @@
-package util;
-
-import util.Link;
+package utils;
 
 import java.io.BufferedReader;
 import java.io.File;
