+++ src/main/java/utils/HTML.java
@@ -1,4 +1,4 @@
-package util;
+package utils;
 
 import java.util.Objects;
